Implemented two first tests + Fixed global player data store not properly loaded
Former-commit-id: 8375d2050e8d0c28f7009d22d03c9d93d1a24610

+++ src/playerbot/RandomPlayerbotMgr.cpp
@@ -530,7 +530,7 @@ vector<uint32> RandomPlayerbotMgr::GetFreeBots(bool alliance)
         do
         {
             Field* fields = results->Fetch();
-            uint32 bot = fields[0].GetUInt32();
+            uint32 bot = uint64(fields[0].GetUInt64());
             bots.insert(bot);
         } while (results->NextRow());
     }
